pyproject.toml
[project]
name = "fqagent"
version = "0.1.0"
description = "芳谦未来智能项目 — 多角色 Agent 编排"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
